chore(suite): rework DeviceWarning component
