+++ packages/suite/src/views/suite/SwitchDevice/CardWithDevice.tsx
@@ -4,9 +4,13 @@ import { AnimatePresence, motion } from 'framer-motion';
 import styled from 'styled-components';
 
 import * as deviceUtils from '@suite-common/suite-utils';
-import { Card, IconName, motionAnimation, useElevation } from '@trezor/components';
+import { acquireDevice, selectDeviceThunk } from '@suite-common/wallet-core';
+import { Banner, Card, IconName, motionAnimation, useElevation } from '@trezor/components';
 import { Elevation, mapElevationToBorder, spacingsPx } from '@trezor/theme';
 
+import { redirectAfterWalletSelectedThunk } from 'src/actions/wallet/addWalletThunk';
+import { Translation } from 'src/components/suite';
+import { useDevice, useDispatch } from 'src/hooks/suite';
 import type { ForegroundAppProps, TrezorDevice } from 'src/types/suite';
 
 import { DeviceHeader } from './DeviceItem/DeviceHeader';
@@ -39,7 +43,6 @@ const DeviceWrapper = styled.div`
 
 interface CardWithDeviceProps {
     children: ReactNode;
-    deviceWarning?: ReactNode;
     onCancel?: ForegroundAppProps['onCancel'];
     device: TrezorDevice;
     isFullHeaderVisible?: boolean;
@@ -52,7 +55,6 @@ export const CardWithDevice = ({
     children,
     onCancel,
     device,
-    deviceWarning,
     isFullHeaderVisible = false,
     onBackButtonClick,
     isFindTrezorVisible,
@@ -63,6 +65,24 @@ export const CardWithDevice = ({
     const needsAttention = deviceUtils.deviceNeedsAttention(deviceStatus);
     const isUnknown = device.type !== 'acquired';
     const { elevation } = useElevation();
+
+    const deviceStatusMessage = deviceUtils.getDeviceNeedsAttentionMessage(deviceStatus);
+    const isLocked = useDevice().isLocked(true);
+    const dispatch = useDispatch();
+
+    const onSolveIssueClick = () => {
+        const needsAcquire =
+            device.type === 'unacquired' ||
+            deviceStatus === 'used-in-other-window' ||
+            deviceStatus === 'was-used-in-other-window';
+        if (needsAcquire) {
+            dispatch(acquireDevice(device));
+        } else {
+            dispatch(selectDeviceThunk({ device }));
+            dispatch(redirectAfterWalletSelectedThunk());
+            onCancel?.(false);
+        }
+    };
 
     return (
         <Card paddingType="small">
@@ -77,7 +97,22 @@ export const CardWithDevice = ({
                     icon={icon}
                 />
 
-                {deviceWarning}
+                {needsAttention && (
+                    <Banner
+                        variant="warning"
+                        rightContent={
+                            <Banner.Button
+                                onClick={onSolveIssueClick}
+                                data-testid="@switch-device/solve-issue-button"
+                                isDisabled={isLocked}
+                            >
+                                <Translation id="TR_SOLVE_ISSUE" />
+                            </Banner.Button>
+                        }
+                    >
+                        {deviceStatusMessage && <Translation id={deviceStatusMessage} />}
+                    </Banner>
+                )}
 
                 {!needsAttention && (
                     <AnimatePresence initial={false}>
